requirements.txt
pandas==1.1.4
streamlit==0.71.0
matplotlib==3.3.3
SQLAlchemy==1.3.20
seaborn==0.11.0
numpy==1.19.4
requests==2.25.0
python-dotenv==0.15.0
scikit_learn==0.23.2
psycopg2-binary==2.8.6
fbprophet==0.7.1
dill==0.3.3